- models updates and fixes

+++ sources/dronesapp-models/src/main/java/com/drones4hire/dronesapp/models/db/commons/Duration.java
@@ -1,46 +1,53 @@
 package com.drones4hire.dronesapp.models.db.commons;
 
-import java.math.BigDecimal;
-
 import com.drones4hire.dronesapp.models.db.AbstractEntity;
 
-public class Duration extends AbstractEntity {
+public class Duration extends AbstractEntity
+{
 	private static final long serialVersionUID = -8574854731176821945L;
 
 	private String title;
-	private BigDecimal min;
-	private BigDecimal max;
+	private Double min;
+	private Double max;
 	private Integer order;
 
-	public String getTitle() {
+	public String getTitle()
+	{
 		return title;
 	}
 
-	public void setTitle(String title) {
+	public void setTitle(String title)
+	{
 		this.title = title;
 	}
 
-	public BigDecimal getMin() {
+	public Double getMin()
+	{
 		return min;
 	}
 
-	public void setMin(BigDecimal min) {
+	public void setMin(Double min)
+	{
 		this.min = min;
 	}
 
-	public BigDecimal getMax() {
+	public Double getMax()
+	{
 		return max;
 	}
 
-	public void setMax(BigDecimal max) {
+	public void setMax(Double max)
+	{
 		this.max = max;
 	}
 
-	public Integer getOrder() {
+	public Integer getOrder()
+	{
 		return order;
 	}
 
-	public void setOrder(Integer order) {
+	public void setOrder(Integer order)
+	{
 		this.order = order;
 	}
 }
